+++ packages/sources/test/orm/go/names.go
@@ -1,0 +1,5 @@
+package model
+
+func (Repo) TableName() string {
+	return "repos"
+}